Added json library, utils and threads to integrate the microbit simulator code
Tested threads - sort of - still get blocking of the code thread, reason
uknown. However it looks like as it is to do with STDOUT as the code
runs fine with you pipe the output to a file. Seems to be a conflict
with epoll_wait in some way.

Removed servuinoFunc, now need to improve the pin states and handle the
pipe events.